EA-43/44/45: Modifications based on feedback given
