+++ condition-list/src/main/java/org/openmrs/Condition.java
@@ -11,14 +11,11 @@
  *
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
-package org.openmrs.module.emrapi.conditionlist.domain;
+package org.openmrs;
 
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.openmrs.BaseOpenmrsData;
-import org.openmrs.Concept;
-import org.openmrs.Patient;
 import org.simpleframework.xml.Attribute;
 
 import java.util.Date;
