+++ condition-list/src/test/java/org/openmrs/dao/HibernateConditionDAOIT.java
@@ -12,7 +12,7 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.module.emrapi.conditionlist.dao;
+package org.openmrs.dao;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -20,7 +20,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.openmrs.Patient;
 import org.openmrs.api.PatientService;
-import org.openmrs.module.emrapi.conditionlist.domain.Condition;
+import org.openmrs.Condition;
 import org.openmrs.test.BaseModuleContextSensitiveTest;
 import org.springframework.beans.factory.annotation.Autowired;
